fix to swallowed exception message
In the `except` block the variable `place_order_result` does not get set if the key `result` is not found in the `place_order_load` dictionary (i.e. the scenario when no result is obtained from the http request). As such, the error message is not in the `place_order_result` dictionary.